+++ src/constants.js
@@ -1,3 +1,5 @@
+export const CONFIG_FILE_NAME = 'nwb.config.js'
+
 export const REACT_VERSION = '15.x'
 
 export const REACT_APP = 'react-app'