+++ templates/react-component/nwb.config.js
@@ -1,11 +1,3 @@
 module.exports = {
-  type: 'react-component',
-  build: {
-    externals: {
-      'react': 'React'
-    },
-    global: '{{globalVariable}}',
-    jsNext: {{jsNext}},
-    umd: {{umd}}
-  }
+  type: 'react-component'
 }