+++ templates/web-module/nwb.config.js
@@ -1,9 +1,3 @@
 module.exports = {
-  type: 'web-module',
-  build: {
-    externals: {},
-    global: '{{globalVariable}}',
-    jsNext: {{jsNext}},
-    umd: {{umd}},
-  }
+  type: 'web-module'
 }